+++ code/src/java/pcgen/cdom/base/ConcreteTransitionChoice.java
@@ -196,7 +196,7 @@ public class ConcreteTransitionChoice<T> implements TransitionChoice<T>
 		{
 			if (choiceActor == null || choiceActor.allow(item, pc, allowStack))
 			{
-				if (assocList != null && stackLimit != null && stackLimit.intValue() > 0)
+				if (assocList != null && stackLimit != null && stackLimit > 0)
 				{
 					int takenCount = 0;
 					for (Object choice : assocList)
@@ -206,7 +206,7 @@ public class ConcreteTransitionChoice<T> implements TransitionChoice<T>
 							takenCount++;
 						}
 					}
-					if (stackLimit.intValue() <= takenCount)
+					if (stackLimit <= takenCount)
 					{
 						continue;
 					}
@@ -330,7 +330,7 @@ public class ConcreteTransitionChoice<T> implements TransitionChoice<T>
 	@Override
 	public void setStackLimit(int limit)
 	{
-		stackLimit = Integer.valueOf(limit);
+		stackLimit = limit;
 	}
 
 	/**
