+++ code/src/java/pcgen/cdom/facet/BonusSkillRankChangeFacet.java
@@ -74,7 +74,7 @@ public class BonusSkillRankChangeFacet extends AbstractStorageFacet<CharID>
 			}
 		
 			Double oldValue = map.get(s);
-			if ((oldValue == null) || (newValue != oldValue.doubleValue()))
+			if ((oldValue == null) || (newValue != oldValue))
 			{
 				map.put(s, newValue);
 				support.fireSkillRankChange(id, s, oldValue, newValue);
@@ -389,7 +389,7 @@ public class BonusSkillRankChangeFacet extends AbstractStorageFacet<CharID>
 			Double rank = map.get(skill);
 			if (rank != null) 
 			{
-				return rank.doubleValue();
+				return rank;
 			}
 		}
 		return 0.0;
