+++ code/src/java/pcgen/cdom/processor/HitDieFormula.java
@@ -66,8 +66,7 @@ public class HitDieFormula implements Processor<HitDie>
 	@Override
 	public HitDie applyProcessor(HitDie origHD, Object context)
 	{
-		return new HitDie(formula.resolve(Integer.valueOf(origHD.getDie()))
-				.intValue());
+		return new HitDie(formula.resolve(origHD.getDie()));
 	}
 
 	/**
