+++ code/src/java/pcgen/core/GameMode.java
@@ -1742,7 +1742,7 @@ public final class GameMode implements Comparable<Object>, GameModeFacade
 			pointBuyStatCosts = new TreeMap<>(new ComparableComparator<>());
 		}
 		abilityScoreCost = null;
-		pointBuyStatCosts.put(Integer.valueOf(pbc.getStatValue()), pbc);
+		pointBuyStatCosts.put(pbc.getStatValue(), pbc);
 	}
 
 	/**
@@ -1886,7 +1886,7 @@ public final class GameMode implements Comparable<Object>, GameModeFacade
 				final PointBuyCost pbc = pointBuyStatCosts.get(statValue);
 				if (pbc.qualifies(pc, null))
 				{
-					lastStat = statValue.intValue();
+					lastStat = statValue;
 				}
 			}
 		}
