+++ code/src/java/pcgen/core/system/LoadInfo.java
@@ -311,7 +311,7 @@ public class LoadInfo implements Loadable
 		 */
 		public int getCheckPenalty()
 		{
-			return checkPenalty.intValue();
+			return checkPenalty;
 		}
 	}
