+++ code/src/java/pcgen/io/exporttoken/EqToken.java
@@ -317,7 +317,7 @@ public class EqToken extends Token
 			//Need a special breakout based on PlayerCharacter reset of IntegerKey.RANGE
 			if (eq.isType("Both") && eq.isType("Melee"))
 			{
-				return Integer.valueOf(0);
+				return 0;
 			}
 			return ((Number) eq.getLocalVariable(pc.getCharID(), rangeVar)).intValue();
 		}
@@ -441,7 +441,7 @@ public class EqToken extends Token
 	 */
 	public static int getAcModTokenInt(PlayerCharacter pc, Equipment eq)
 	{
-		return eq.getACMod(pc).intValue();
+		return eq.getACMod(pc);
 	}
 
 	/**
@@ -528,7 +528,7 @@ public class EqToken extends Token
 	 */
 	public static float getCarriedTokenFloat(Equipment eq)
 	{
-		return eq.numberCarried().floatValue();
+		return eq.numberCarried();
 	}
 
 	/**
@@ -996,7 +996,7 @@ public class EqToken extends Token
 	public static String getRangeToken(Equipment eq, PlayerCharacter pc)
 	{
 		return Globals.getGameModeUnitSet().displayDistanceInUnitSet(
-			getRange(pc, eq).intValue())
+				getRange(pc, eq))
 			+ Globals.getGameModeUnitSet().getDistanceUnit();
 	}
