+++ code/src/java/plugin/exporttokens/SkillSitToken.java
@@ -285,7 +285,7 @@ public class SkillSitToken extends Token
 
 				case SkillToken.SKILL_TOTAL:
 					int rank = SkillRankControl.getTotalRank(pc, skill).intValue()
-						+ SkillModifier.modifier(skill, pc).intValue();
+						+ SkillModifier.modifier(skill, pc);
 					if (isSituation)
 					{
 						rank += sit.getSituationBonus();
@@ -315,7 +315,7 @@ public class SkillSitToken extends Token
 					break;
 
 				case SkillToken.SKILL_MOD:
-					int mod = SkillModifier.modifier(skill, pc).intValue();
+					int mod = SkillModifier.modifier(skill, pc);
 					if (isSituation)
 					{
 						mod += sit.getSituationBonus();
@@ -332,7 +332,7 @@ public class SkillSitToken extends Token
 					break;
 
 				case SkillToken.SKILL_MISC:
-					int misc = SkillModifier.modifier(skill, pc).intValue();
+					int misc = SkillModifier.modifier(skill, pc);
 					if (isSituation)
 					{
 						misc += sit.getSituationBonus();
@@ -394,8 +394,7 @@ public class SkillSitToken extends Token
 					{
 						int mRank =
 								etRank
-									+ SkillModifier.modifier(skill, pc)
-										.intValue();
+									+ SkillModifier.modifier(skill, pc);
 						if (isSituation)
 						{
 							mRank += sit.getSituationBonus();
@@ -416,8 +415,7 @@ public class SkillSitToken extends Token
 					{
 						int mRank =
 								tRank
-									+ SkillModifier.modifier(skill, pc)
-										.intValue();
+									+ SkillModifier.modifier(skill, pc);
 						if (isSituation)
 						{
 							mRank += sit.getSituationBonus();
