+++ code/src/java/plugin/initiative/gui/StartEvent.java
@@ -99,7 +99,7 @@ public class StartEvent extends javax.swing.JDialog
 		this.initiative = initiative;
 		tName.grabFocus();
 		tPlayer.setText(player);
-		lInit.setValue(Integer.valueOf(init));
+		lInit.setValue(init);
 		initCheckBox();
 	}
 
@@ -151,8 +151,8 @@ public class StartEvent extends javax.swing.JDialog
 	protected void save()
 	{
 		initiative.initList.add(new Event(tName.getText(), tPlayer.getText(),
-			tEffect.getText(), ((Integer) lDuration.getValue()).intValue(),
-			((Integer) lInit.getValue()).intValue(), cbAlert.isSelected()));
+			tEffect.getText(), (Integer) lDuration.getValue(),
+				(Integer) lInit.getValue(), cbAlert.isSelected()));
 		initiative.writeToCombatTabWithRound(" Event Timer " + tName.getText()
 			+ " Started");
 		initiative.refreshTable();
@@ -282,7 +282,7 @@ public class StartEvent extends javax.swing.JDialog
 		initiativeLabel = new javax.swing.JLabel();
 		sInit = Utils.buildSlider(1, 50);
 		lInit = Utils.buildIntegerFieldWithSlider(sInit);
-		lInit.setValue(Integer.valueOf(1));
+		lInit.setValue(1);
 
 		java.awt.GridBagConstraints gridBagConstraints;
 		initiativeLabel.setText("Initiative");
@@ -318,7 +318,7 @@ public class StartEvent extends javax.swing.JDialog
 		durationLabel = new javax.swing.JLabel();
 		sDuration = Utils.buildSlider(1, 50);
 		lDuration = Utils.buildIntegerFieldWithSlider(sDuration);
-		lDuration.setValue(Integer.valueOf(1));
+		lDuration.setValue(1);
 
 		java.awt.GridBagConstraints gridBagConstraints;
 		durationLabel.setText("Duration");
